feat: добавлены новые команды для Telegram бота, улучшена обработка сообщений и добавлена поддержка статусов задач; обновлены маршруты и компоненты для улучшения пользовательского интерфейса

+++ back/src/entities/ChatRole.ts
@@ -7,7 +7,7 @@ export class ChatRole {
   @PrimaryGeneratedColumn('increment')
   id!: number;
 
-  @Column({ name: 'chat_id', type: 'int' })
+  @Column({ name: 'chat_id', type: 'bigint' })
   chatId!: number;
 
   @Column({ name: 'role_id', type: 'int' })
